pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "histo-omics-lite"
version = "0.1.0"
description = "Minimal synthetic histology-omics alignment package"
readme = "README.md"
authors = [{ name = "Altalanta", email = "engineering@altalanta.ai" }]
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
dependencies = [
  "torch>=2.1,<2.3",
  "torchvision>=0.16,<0.19",
  "lightning>=2.2,<3.0",
  "hydra-core>=1.3,<1.4",
  "pydantic>=1.10,<2.0",
  "numpy>=1.24,<2.0",
  "pillow>=9.0,<11.0",
  "scikit-learn>=1.3,<2.0",
  "matplotlib>=3.5,<4.0",
  "pandas>=1.5,<3.0"
]

[project.urls]
Homepage = "https://github.com/altalanta/histo-omics-lite"
Issues = "https://github.com/altalanta/histo-omics-lite/issues"

[project.optional-dependencies]
cpu = [
  "mlflow>=2.8,<3.0",
  "jinja2>=3.1,<4.0",
  "umap-learn>=0.5,<1.0"
]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "mypy>=1.8",
  "ruff>=0.1.14",
  "pre-commit>=3.6",
  "build>=1.0",
  "mlflow>=2.8,<3.0",
  "jinja2>=3.1,<4.0",
  "umap-learn>=0.5,<1.0"
]

[project.scripts]
histo-omics-lite = "histo_omics_lite.inference.cli:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "PL", "UP", "B", "A", "C4"]
ignore = ["E501", "PLR2004", "PLR0913", "N812"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.pytest.ini_options]
addopts = "-ra -q --cov=histo_omics_lite --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["histo_omics_lite"]
parallel = false
omit = []

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
strict = true
packages = ["histo_omics_lite"]
